# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "N100 Financial Intelligence Platform",
    "description": "FastAPI Backend for N100 Financial Intelligence Platform",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/companies": {
      "get": {
        "tags": [
          "Companies"
        ],
        "summary": "Get List of Companies",
        "description": "Retrieve list of all authoritative companies with optional sector, market cap, and search filters.",
        "operationId": "get_companies_api_v1_companies_get",
        "parameters": [
          {
            "name": "sector",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by broad or sub sector (case-insensitive)",
              "title": "Sector"
            },
            "description": "Filter by broad or sub sector (case-insensitive)"
          },
          {
            "name": "market_cap_category",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by market cap category (e.g. Large Cap)",
              "title": "Market Cap Category"
            },
            "description": "Filter by market cap category (e.g. Large Cap)"
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Partial search by company name or ticker",
              "title": "Search"
            },
            "description": "Partial search by company name or ticker"
          }
        ],
        "responses": {
          "200": {
            "description": "List of companies returned successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CompanyListItem"
                  },
                  "title": "Response Get Companies Api V1 Companies Get"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/screener": {
      "get": {
        "tags": [
          "Screener"
        ],
        "summary": "Screen Companies by Financial KPIs",
        "description": "Filter Nifty 100 companies based on ROE, D/E, Free Cash Flow, Sector, CAGR, and P/E ratio.",
        "operationId": "get_screener_results_api_v1_screener_get",
        "parameters": [
          {
            "name": "min_roe",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Minimum Return on Equity % (ROE >= threshold)",
              "title": "Min Roe"
            },
            "description": "Minimum Return on Equity % (ROE >= threshold)"
          },
          {
            "name": "max_de",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Maximum Debt to Equity ratio (D/E <= threshold)",
              "title": "Max De"
            },
            "description": "Maximum Debt to Equity ratio (D/E <= threshold)"
          },
          {
            "name": "min_fcf",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Minimum Free Cash Flow (FCF >= threshold)",
              "title": "Min Fcf"
            },
            "description": "Minimum Free Cash Flow (FCF >= threshold)"
          },
          {
            "name": "sector",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Exact or partial sector filter",
              "title": "Sector"
            },
            "description": "Exact or partial sector filter"
          },
          {
            "name": "min_rev_cagr_5yr",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Minimum 5-Year Revenue CAGR %",
              "title": "Min Rev Cagr 5Yr"
            },
            "description": "Minimum 5-Year Revenue CAGR %"
          },
          {
            "name": "min_pat_cagr_5yr",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Minimum 5-Year PAT CAGR %",
              "title": "Min Pat Cagr 5Yr"
            },
            "description": "Minimum 5-Year PAT CAGR %"
          },
          {
            "name": "max_pe",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Maximum Price to Earnings ratio (P/E <= threshold)",
              "title": "Max Pe"
            },
            "description": "Maximum Price to Earnings ratio (P/E <= threshold)"
          }
        ],
        "responses": {
          "200": {
            "description": "Filtered and ranked company list returned successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ScreenerCompanyItem"
                  },
                  "title": "Response Get Screener Results Api V1 Screener Get"
                }
              }
            }
          },
          "400": {
            "description": "Invalid query parameters"
          },
          "500": {
            "description": "Internal server error"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sectors": {
      "get": {
        "tags": [
          "Sectors"
        ],
        "summary": "Get Sector Summaries",
        "description": "Retrieve sector aggregated statistics including company count, median ROE, median P/E, and median D/E.",
        "operationId": "get_sectors_api_v1_sectors_get",
        "responses": {
          "200": {
            "description": "Sector summary list returned successfully",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SectorSummaryItem"
                  },
                  "type": "array",
                  "title": "Response Get Sectors Api V1 Sectors Get"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/v1/sectors/{sector}/companies": {
      "get": {
        "tags": [
          "Sectors"
        ],
        "summary": "Get Companies by Sector",
        "description": "Retrieve list of companies and latest financial KPIs belonging to a specific sector.",
        "operationId": "get_sector_companies_api_v1_sectors__sector__companies_get",
        "parameters": [
          {
            "name": "sector",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Sector"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of companies in sector returned successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SectorCompanyItem"
                  },
                  "title": "Response Get Sector Companies Api V1 Sectors  Sector  Companies Get"
                }
              }
            }
          },
          "404": {
            "description": "Sector not found"
          },
          "500": {
            "description": "Internal server error"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/peers/{group_name}": {
      "get": {
        "tags": [
          "Peers"
        ],
        "summary": "Get Peer Group Companies and Percentiles",
        "description": "Retrieve percentile rankings for all 10 core peer metrics across companies in a peer group.",
        "operationId": "get_peer_group_details_api_v1_peers__group_name__get",
        "parameters": [
          {
            "name": "group_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Group Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Peer group percentile data returned successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PeerGroupResponse"
                }
              }
            }
          },
          "404": {
            "description": "Peer group not found"
          },
          "500": {
            "description": "Internal server error"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/companies/{ticker}/peers/compare": {
      "get": {
        "tags": [
          "Peers"
        ],
        "summary": "Get Peer Radar Comparison Data",
        "description": "Retrieve 8 radar axis metrics for requested company, peer group average, and benchmark company.",
        "operationId": "get_company_peer_compare_api_v1_companies__ticker__peers_compare_get",
        "parameters": [
          {
            "name": "ticker",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ticker"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Radar comparison data returned successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RadarCompareResponse"
                }
              }
            }
          },
          "404": {
            "description": "Company ticker not found"
          },
          "500": {
            "description": "Internal server error"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/market-cap/{ticker}": {
      "get": {
        "tags": [
          "Valuation"
        ],
        "summary": "Get Historical Valuation Multiples",
        "description": "Retrieve historical valuation multiples (P/E, P/B, EV/EBITDA, Dividend Yield) from 2019 to 2024 for a company.",
        "operationId": "get_company_market_cap_valuation_api_v1_market_cap__ticker__get",
        "parameters": [
          {
            "name": "ticker",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ticker"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Historical valuation data returned successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompanyValuationResponse"
                }
              }
            }
          },
          "404": {
            "description": "Company ticker not found"
          },
          "500": {
            "description": "Internal server error"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/portfolio/stats": {
      "get": {
        "tags": [
          "Portfolio"
        ],
        "summary": "Get Portfolio KPI Percentile Statistics",
        "description": "Retrieve statistical percentiles (P10, P25, P50, P75, P90, Mean, Std) for core financial KPIs across the company universe.",
        "operationId": "get_portfolio_stats_api_v1_portfolio_stats_get",
        "responses": {
          "200": {
            "description": "Portfolio statistics returned successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PortfolioStatsResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/v1/companies/{ticker}/documents": {
      "get": {
        "tags": [
          "Documents"
        ],
        "summary": "Get Company Annual Reports & Documents",
        "description": "Retrieve annual report links and link validity status for requested company.",
        "operationId": "get_company_documents_api_v1_companies__ticker__documents_get",
        "parameters": [
          {
            "name": "ticker",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ticker"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Document list returned successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompanyDocumentsResponse"
                }
              }
            }
          },
          "404": {
            "description": "Company ticker not found"
          },
          "500": {
            "description": "Internal server error"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "CompanyListItem": {
        "properties": {
          "company_id": {
            "type": "string",
            "title": "Company Id",
            "description": "Unique company ticker identifier"
          },
          "company_name": {
            "type": "string",
            "title": "Company Name",
            "description": "Official company name"
          },
          "broad_sector": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Broad Sector",
            "description": "Broad industry sector classification"
          },
          "sub_sector": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sub Sector",
            "description": "Specific sub-sector classification"
          },
          "market_cap_category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Market Cap Category",
            "description": "Market cap category (e.g. Large Cap)"
          },
          "roe_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Roe Pct",
            "description": "Return on Equity percentage"
          },
          "roce_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Roce Pct",
            "description": "Return on Capital Employed percentage"
          }
        },
        "type": "object",
        "required": [
          "company_id",
          "company_name"
        ],
        "title": "CompanyListItem"
      },
...
      "CompanyValuationResponse": {
        "properties": {
          "ticker": {
            "type": "string",
            "title": "Ticker",
            "description": "Company ticker symbol"
          },
          "company_name": {
            "type": "string",
            "title": "Company Name",
            "description": "Official company name"
          },
          "historical_valuation": {
            "items": {
              "$ref": "#/components/schemas/ValuationYearRecord"
            },
            "type": "array",
            "title": "Historical Valuation",
            "description": "Chronological annual valuation records (2019 to 2024)"
          }
        },
        "type": "object",
        "required": [
          "ticker",
          "company_name",
          "historical_valuation"
        ],
        "title": "CompanyValuationResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PeerGroupResponse": {
        "properties": {
          "peer_group_name": {
            "type": "string",
            "title": "Peer Group Name",
            "description": "Peer group identifier"
          },
          "company_count": {
            "type": "integer",
            "title": "Company Count",
            "description": "Total companies in peer group"
          },
          "companies": {
            "items": {
              "$ref": "#/components/schemas/PeerCompanyPercentileItem"
            },
            "type": "array",
            "title": "Companies",
            "description": "List of companies with 10 metric percentiles"
          }
        },
        "type": "object",
        "required": [
          "peer_group_name",
          "company_count",
          "companies"
        ],
        "title": "PeerGroupResponse"
      },
...
      "ScreenerCompanyItem": {
        "properties": {
          "company_id": {
            "type": "string",
            "title": "Company Id",
            "description": "Unique company ticker symbol"
          },
          "company_name": {
            "type": "string",
            "title": "Company Name",
            "description": "Official company name"
          },
          "sector": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sector",
            "description": "Industry sector classification"
          },
          "rank": {
            "type": "integer",
            "title": "Rank",
            "description": "Screener ranking position"
          },
          "ranking": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ranking",
            "description": "Alias for screener ranking position"
          },
          "roe": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Roe",
            "description": "Return on Equity %"
          },
          "debt_to_equity": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Debt To Equity",
            "description": "Debt to Equity ratio"
          },
          "de": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "De",
            "description": "Alias for Debt to Equity ratio"
          },
          "free_cash_flow": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Free Cash Flow",
            "description": "Free Cash Flow in Crores/Millions"
          },
          "fcf": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fcf",
            "description": "Alias for Free Cash Flow"
          },
          "revenue_cagr_5yr": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Revenue Cagr 5Yr",
            "description": "5-Year Revenue CAGR %"
          },
          "pat_cagr_5yr": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pat Cagr 5Yr",
            "description": "5-Year Net Profit CAGR %"
          },
          "pe_ratio": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pe Ratio",
            "description": "Price to Earnings ratio"
          },
          "pe": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pe",
            "description": "Alias for Price to Earnings ratio"
          }
        },
        "type": "object",
        "required": [
          "company_id",
          "company_name",
          "rank"
        ],
        "title": "ScreenerCompanyItem"
      },
      "SectorCompanyItem": {
        "properties": {
          "company_id": {
            "type": "string",
            "title": "Company Id",
            "description": "Unique company ticker symbol"
          },
          "company_name": {
            "type": "string",
            "title": "Company Name",
            "description": "Official company name"
          },
          "sector": {
            "type": "string",
            "title": "Sector",
            "description": "Sector classification"
          },
          "roe": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Roe",
            "description": "Latest Return on Equity %"
          },
          "roce": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Roce",
            "description": "Latest Return on Capital Employed %"
          },
          "debt_to_equity": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Debt To Equity",
            "description": "Latest Debt to Equity ratio"
          },
          "pe_ratio": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pe Ratio",
            "description": "Latest Price to Earnings ratio"
          },
          "net_profit_margin": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Net Profit Margin",
            "description": "Latest Net Profit Margin %"
          },
          "latest_kpis": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest Kpis",
            "description": "Consolidated latest financial KPIs"
          }
        },
        "type": "object",
        "required": [
          "company_id",
          "company_name",
          "sector"
        ],
        "title": "SectorCompanyItem"
      },
      "SectorSummaryItem": {
        "properties": {
          "sector": {
            "type": "string",
            "title": "Sector",
            "description": "Name of the industry sector"
          },
          "company_count": {
            "type": "integer",
            "title": "Company Count",
            "description": "Number of companies in this sector"
          },
          "median_roe": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Median Roe",
            "description": "Median Return on Equity % across sector"
          },
          "median_pe": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Median Pe",
            "description": "Median Price to Earnings ratio across sector"
          },
          "median_de": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Median De",
            "description": "Median Debt to Equity ratio across sector"
          }
        },
        "type": "object",
        "required": [
          "sector",
          "company_count"
        ],
        "title": "SectorSummaryItem"
      },
...
    }
  }
}